+++ app.go
@@ -20,7 +20,7 @@ import (
 	"time"
 )
 
-var version = "0.6.1"
+var version = "0.7.1"
 
 const (
 	maxCachedQueries = 512